+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - The place_sampling_points() function to calculate the sampling points coordinates in the roi_align.py file.
 - The bilinear_interpolate() function to perform bilinear interpolation on all sampling points in the roi_align.py file.
+- The pool_interpolations() function to pool the bilinear interpolations in the roi_align.py file.
 - The display_aligned_rois() function to visualize a sample of the aligned rois in the visualize.py file.
 
 ### Changed